sorts timestamped photos into folders
this is for an alternative timelapse script that adds a timestamp to the photos and sorts them into folders by year/month/day.  

Note this is not for use with the default pipic.sh script as this can't guarantee a wifi connection for setting the date, and hence this used an incremental numbering system.