presentation slides from Ogcamp 2017
I gave a 25 min talk about my timelapse tinkering at Oggcamp 2017, and these are the slides.  

Unfortunately for anyone wanting to relive my presentation I'm very critical of those which have lots of words for the audience to read, as it detracts from the words being said by the presenter.  Hence there is very little readable information in these, for which I don't apologise.